+++ test/unit/specs/lcdClient.spec.js
@@ -15,6 +15,8 @@ const mung = require(`express-mung`)
 const http = require(`http`)
 let LcdClient = require(`renderer/connectors/lcdClient.js`)
 let lcdClientMock = require(`renderer/connectors/lcdClientMock.js`)
+let { proposals, deposits, votes } = lcdClientMock.state
+
 const path = require(`path`)
 const createMiddleware = require(`swagger-express-middleware`)
 const { MemoryDataStore, Resource } = createMiddleware
@@ -474,5 +476,94 @@ describe(`LCD Client`, () => {
         undefined
       ])
     })
+  })
+
+  /* ============ Governance ============ */
+
+  it(`fetches all governance proposals`, async () => {
+    axios.get = jest.fn().mockReturnValue({})
+    await client.queryProposals()
+    expect(axios.get.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals`,
+      undefined
+    ])
+  })
+
+  it(`queries a single proposal`, async () => {
+    axios.get = jest.fn().mockReturnValue({})
+    await client.queryProposal(1)
+    expect(axios.get.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals/1`,
+      undefined
+    ])
+  })
+
+  it(`queries a proposal votes`, async () => {
+    axios.get = jest.fn().mockReturnValue({})
+    await client.queryProposalVotes(1)
+    expect(axios.get.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals/1/votes`,
+      undefined
+    ])
+  })
+
+  it(`queries a proposal vote from an address`, async () => {
+    axios.get = jest.fn().mockReturnValue({})
+    await client.queryProposalVote(
+      1,
+      `cosmos15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9`
+    )
+    expect(axios.get.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals/1/votes/cosmos15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9`,
+      undefined
+    ])
+  })
+
+  it(`queries a proposal deposits`, async () => {
+    axios.get = jest.fn().mockReturnValue({})
+    await client.queryProposalDeposits(1)
+    expect(axios.get.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals/1/deposits`,
+      undefined
+    ])
+  })
+
+  it(`queries a proposal deposit from an address`, async () => {
+    axios.get = jest.fn().mockReturnValue({})
+    await client.queryProposalDeposit(
+      1,
+      `cosmos15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9`
+    )
+    expect(axios.get.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals/1/deposits/cosmos15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9`,
+      undefined
+    ])
+  })
+
+  it(`submits a new proposal`, async () => {
+    axios.post = jest.fn().mockReturnValue({})
+    await client.submitProposal(proposals[0])
+    expect(axios.post.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals`,
+      proposals[0]
+    ])
+  })
+
+  it(`submits a new vote to a proposal`, async () => {
+    axios.post = jest.fn().mockReturnValue({})
+    await client.submitVote(proposals[0].proposal_id, votes[0])
+    expect(axios.post.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals/${proposals[0].proposal_id}/votes`,
+      votes[0]
+    ])
+  })
+
+  it(`submits a new deposit to a proposal`, async () => {
+    axios.post = jest.fn().mockReturnValue({})
+    await client.submitDeposit(proposals[0].proposal_id, deposits[0])
+    expect(axios.post.mock.calls[0]).toEqual([
+      `${remoteLcdURL}/gov/proposals/${proposals[0].proposal_id}/deposits`,
+      deposits[0]
+    ])
   })
 })