+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Contract tests for `keys` endpoints in lcdClient. @NodeGuy
 * Protected Gaia Lite process with a mutex to prevent locking error. @NodeGuy
 * catching errors on cache encryption. @faboweb
+* #1436 governance endpoints and vuex module @fedekunze
 
 ### Changed